+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+# [2.14.0] - 2018-01-10
+## Added
+* Allow reading/writing `Source.canonical` at `/api/v2/sources/`
+* Include `<author>` in atom feed at `/api/v2/atom/`
+* ScholarsArchive@OSU source config for their new API
+
+## Changed
+* Prevent OSF harvester from being throttled
+* Update NSFAwards harvester/transformer to include more fields
+
 # [2.13.1] - 2018-01-04
 ## Fixed
 * Use request context to build URLs in the API instead of SHARE_API_URL setting